+++ server/lib/event/get-tag-event-traits.js
@@ -6,9 +6,13 @@ export default function getTagEventTraits(ctx: Object, user: Object, allTags: Ar
     if (!t.name) {
       t = _.find(allTags, { id: t.id });
     }
+    if (!t.name) {
+      ctx.client.logger.debug("incoming.event.tagNotFound");
+      return "";
+    }
     return t.name;
   });
   const traits = {};
-  traits["intercom/tags"] = tags;
+  traits["intercom/tags"] = _.compact(tags);
   return traits;
 }